// frontend/events/src/app/models/user.ts
export class User {
    organization: string;
    street: string;
    number: string;
    city: string;
    first_name: string;
    last_name: string;
    telephone: string;
    areaactivity: string;
    website: string;
    email: string;
    username: string;
    password: string;
    organizationInfoText: string;
    privilege: string;
}